% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectDS.R
\name{connectDS}
\alias{connectDS}
\title{Create a connection to DataSpace}
\usage{
connectDS(study = NULL, login = NULL, password = NULL, verbose = FALSE)
}
\arguments{
\item{study}{A character. Name of the study to connect.}

\item{login}{A character. Optional argument. If there is no netrc
file a temporary one can be written by passing login and password of an
active DataSpace account.}

\item{password}{A character. Optional. The password for the selected
login.}

\item{verbose}{A logical. Whether to print the extra details for
troubleshooting.}
}
\value{
an instance of \code{DataSpaceConnection}
}
\description{
Constructor for \code{\link{DataSpaceConnection}}
}
\details{
Instantiates an \code{DataSpaceConnection} for \code{study}
The constructor will try to take the values of the various \code{labkey.*}
parameters from the global environment. If they don't exist, it will use
default values. These are assigned to `options`, which are then used by the
\code{DataSpaceConnection} class.
}
\examples{
\dontrun{
# Single study
con <- connectDS("cvd408")

# Cross study
con <- connectDS("")
}
}
\seealso{
\code{\link{DataSpaceConnection}}
}
