% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConnection.R
\name{CreateConnection}
\alias{CreateConnection}
\title{CreateConnection}
\usage{
CreateConnection(study = NULL, login = NULL, password = NULL,
  verbose = FALSE)
}
\arguments{
\item{study}{A \code{"character"} vector naming the study.}

\item{login}{A \code{"character"}. Optional argument. If there is no netrc
file a temporary one can be written by passing login and password of an
active DataSpace account.}

\item{password}{A \code{"character"}. Optional. The password for the selected
login.}

\item{verbose}{A \code{"logical"} whether to print the extra details for
troubleshooting.}
}
\value{
an instance of an \code{DataSpaceConnection}
}
\description{
CreateConnection
}
\examples{
\dontrun{
# Single study
con <- CreateConnection("cvd408")
# Cross study
con <- CreateConnection("")
}
}
