% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSpaceConnection.R
\docType{class}
\name{DataSpaceConnection-class}
\alias{DataSpaceConnection-class}
\alias{.DSCon}
\title{A Reference Class to represent a bank account.}
\description{
A Reference Class to represent a bank account.
}
\section{Fields}{

\describe{
\item{\code{study}}{A length-one numeric vector.}

\item{\code{config}}{A length-one numeric vector.}

\item{\code{available_datasets}}{A length-one numeric vector.}
}}

\section{Methods}{

\describe{
\item{\code{getAvailableDatasets()}}{Get availabel datasets}

\item{\code{getDataset(dataset_name, original_view = FALSE, colFilter = NULL, ...)}}{get dataset!}

\item{\code{initialize(study = NULL, login = NULL, password = NULL, verbose = FALSE)}}{Initialize class}
}}

