% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSpaceDonorMetadata.R
\name{DataSpaceDonorMetadata}
\alias{DataSpaceDonorMetadata}
\title{The DataSpaceDonorMetadata class}
\description{
The DataSpaceDonorMetadata class

The DataSpaceDonorMetadata class
}
\section{Constructor}{

\code{DataSpaceConnection$getDonorMetadata()}
}

\examples{
\dontrun{
# Create a connection (Initiate a DataSpaceConnection object)
con <- connectDS()

# Get all the donor metadata
donorMeta <- con$getDonorMetadata()

# Load all alignments
donorMeta$loadDaash()

# Or just a subset of the alignments for the metadata extracted so far
donorMeta$loadDaash(donorIds = "cds_donor_32")

# Inspect the `topCalls` field
donorMeta$topCalls

}

}
\seealso{
\code{\link{connectDS}} \code{\link{DataSpaceConnection}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{A}}{data.table. The table of information concerning the runs from the two alignment tools.
measurements against viruses.}

\item{\code{A}}{data.table. The table of allele sequences from reported allele matches.
measurements against viruses.}

\item{\code{A}}{data.table. The table of information concerning the runs from the two alignment tools.
measurements against viruses.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{config}}{A list. Stores configuration of the connection object such
as URL, path and username.}

\item{\code{donorMetadata}}{A data.table. The table of the basic metadata from a donor metadata filter.}

\item{\code{topCalls}}{A data.table. A table of top allele matches from both IgBLAST and V-Quest.}

\item{\code{alignments}}{A data.table. The table of alignments of germline genes to a given sequence.
attributes.}

\item{\code{A}}{data.table. The table of information concerning the runs from the two alignment tools.
measurements against viruses.}

\item{\code{A}}{data.table. The table of allele sequences from reported allele matches.
measurements against viruses.}

\item{\code{A}}{data.table. The table of information concerning the runs from the two alignment tools.
measurements against viruses.}

\item{\code{variableDefinitions}}{A data.table. The table of variable definitions for the metadata.
definitions.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataSpaceDonorMetadata-new}{\code{DataSpaceDonorMetadata$new()}}
\item \href{#method-DataSpaceDonorMetadata-print}{\code{DataSpaceDonorMetadata$print()}}
\item \href{#method-DataSpaceDonorMetadata-refresh}{\code{DataSpaceDonorMetadata$refresh()}}
\item \href{#method-DataSpaceDonorMetadata-loadDaash}{\code{DataSpaceDonorMetadata$loadDaash()}}
\item \href{#method-DataSpaceDonorMetadata-clone}{\code{DataSpaceDonorMetadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceDonorMetadata-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceDonorMetadata-new}{}}}
\subsection{Method \code{new()}}{
Initialize \code{DataSpaceDonorMetadata} object.
See \code{\link{DataSpaceConnection}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSpaceDonorMetadata$new(donorIds, config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{donorIds}}{A character vector of cds donor ids}

\item{\code{config}}{A list.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceDonorMetadata-print"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceDonorMetadata-print}{}}}
\subsection{Method \code{print()}}{
Print the \code{DataSpaceDonor} object summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSpaceDonorMetadata$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceDonorMetadata-refresh"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceDonorMetadata-refresh}{}}}
\subsection{Method \code{refresh()}}{
Refresh the \code{DataSpaceDonorMetadata} object to update datasets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSpaceDonorMetadata$refresh()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceDonorMetadata-loadDaash"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceDonorMetadata-loadDaash}{}}}
\subsection{Method \code{loadDaash()}}{
Applies DAASH tables to the this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSpaceDonorMetadata$loadDaash(donorIds = c(), filter = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceDonorMetadata-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceDonorMetadata-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSpaceDonorMetadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
