% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSpaceMab.R
\docType{class}
\name{DataSpaceMab}
\alias{DataSpaceMab}
\title{The DataSpaceMab class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataSpaceMab
}
\description{
The DataSpaceMab class
}
\section{Constructor}{

\code{DataSpaceConnection$getMab()}
}

\section{Fields}{

\describe{
}
}

\section{Methods}{

\describe{
}
}

\examples{
\dontrun{
}
}
\seealso{
\code{\link{connectDS}} \code{\link{DataSpaceConnection}}
}
\keyword{datasets}
