% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSpaceMabMetadata.R
\name{DataSpaceMabMetadata}
\alias{DataSpaceMabMetadata}
\title{The DataSpaceMabMetadata class}
\description{
The DataSpaceMabMetadata class

The DataSpaceMabMetadata class
}
\section{Constructor}{

\code{DataSpaceConnection$getMabMetadata()}
}

\examples{
\dontrun{
# Create a connection (Initiate a DataSpaceConnection object)
con <- connectDS()

# Browse the mAb Grid
con$mabMetadata

# Filter the grid by viruses
mabMeta <- con$getMabMetadata(mabIds = "cds_mab_32")

# Or get all the mab metadata
mabMeta <- con$getMabMetadata()

# Load all alignments
mabMeta$loadDaash()

# Or just a subset of the alignments for the metadata extracted so far
mabMeta$loadDaash(mabIds = "cds_mab_32")

# Inspect the `topCalls` field
mabMeta$topCalls

}

}
\seealso{
\code{\link{connectDS}} \code{\link{DataSpaceConnection}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{config}}{A list. Stores configuration of the connection object such
as URL, path and username.}

\item{\code{mabMetadata}}{A data.table. The table of the basic metadata from a mab metadata filter.}

\item{\code{mabMix}}{A data.table. The table of mAb mixture IDs and mAb IDs for merging data reported as mAb mixures and mAb metadata.}

\item{\code{topCalls}}{A data.table. A table of top allele matches from both IgBLAST and V-Quest.}

\item{\code{alignments}}{A data.table. The table of alignments of germline genes to a given sequence.
attributes.}

\item{\code{sequences}}{A data.table. The table of information concerning the runs from the two alignment tools.
measurements against viruses.}

\item{\code{alleleSequences}}{A data.table. The table of allele sequences from reported allele matches.
measurements against viruses.}

\item{\code{runInformation}}{A data.table. The table of information concerning the runs from the two alignment tools.
measurements against viruses.}

\item{\code{variableDefinitions}}{A data.table. The table of variable definitions for the metadata.
definitions.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataSpaceMabMetadata-new}{\code{DataSpaceMabMetadata$new()}}
\item \href{#method-DataSpaceMabMetadata-print}{\code{DataSpaceMabMetadata$print()}}
\item \href{#method-DataSpaceMabMetadata-refresh}{\code{DataSpaceMabMetadata$refresh()}}
\item \href{#method-DataSpaceMabMetadata-loadLanlMetadata}{\code{DataSpaceMabMetadata$loadLanlMetadata()}}
\item \href{#method-DataSpaceMabMetadata-getFastaFromSequences}{\code{DataSpaceMabMetadata$getFastaFromSequences()}}
\item \href{#method-DataSpaceMabMetadata-loadDaash}{\code{DataSpaceMabMetadata$loadDaash()}}
\item \href{#method-DataSpaceMabMetadata-clone}{\code{DataSpaceMabMetadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceMabMetadata-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceMabMetadata-new}{}}}
\subsection{Method \code{new()}}{
Initialize \code{DataSpaceMabMetadata} object.
See \code{\link{DataSpaceConnection}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSpaceMabMetadata$new(mabIds, config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mabIds}}{A character vector of cds mab ids.}

\item{\code{config}}{A list.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceMabMetadata-print"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceMabMetadata-print}{}}}
\subsection{Method \code{print()}}{
Print the \code{DataSpaceMab} object summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSpaceMabMetadata$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceMabMetadata-refresh"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceMabMetadata-refresh}{}}}
\subsection{Method \code{refresh()}}{
Refresh the \code{DataSpaceMabMetadata} object to update datasets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSpaceMabMetadata$refresh()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceMabMetadata-loadLanlMetadata"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceMabMetadata-loadLanlMetadata}{}}}
\subsection{Method \code{loadLanlMetadata()}}{
Applies LANL metadata to mabMetadata table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSpaceMabMetadata$loadLanlMetadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceMabMetadata-getFastaFromSequences"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceMabMetadata-getFastaFromSequences}{}}}
\subsection{Method \code{getFastaFromSequences()}}{
Return a fasta file for available daash sequences that have been loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSpaceMabMetadata$getFastaFromSequences(
  seqIds = NULL,
  sequenceType = "nt",
  originalHeaders = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seqIds}}{character CDS Sequence IDs, found in the `sequence_id` column of the donorMetadata table in a DataSpaceDonorMetadata object. These are in the format "cds_seq_###".}

\item{\code{sequenceType}}{character What type of fasta file to return: nt = nucleotide, aa = amino acid.}

\item{\code{originalHeaders}}{boolean T/F if the original fasta headers should be provided, or}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceMabMetadata-loadDaash"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceMabMetadata-loadDaash}{}}}
\subsection{Method \code{loadDaash()}}{
Applies DAASH tables to the this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSpaceMabMetadata$loadDaash(mabIds = c(), filter = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mabIds}}{character CDS MAb IDs, found in the `mab_id` column of the mabMetadata table in a DataSpaceMabMetadata object. These are in the format "cds_mab_###".}

\item{\code{filter}}{character An additional filter to add to the query for getting daash results. This any valid filter for the view `donor_mab_sequence_header_source`, a view found on the DataSpace LabKey server.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceMabMetadata-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceMabMetadata-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSpaceMabMetadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
