% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSpaceMabs.R
\name{DataSpaceMabs}
\alias{DataSpaceMabs}
\title{The DataSpaceMab class}
\description{
An R6 class for DataSpace MAb data.
}
\section{Constructor}{

\code{DataSpaceConnection$getMab()}
}

\examples{
\dontrun{
# Create a connection (Initiate a DataSpaceConnection object)
con <- connectDS()

# Inspect available mabs, then pass subset to the `getMabs` method.
vrc01 <- con$availableMabs[mab_name_std == "VRC01"] |>
  con$getMabs()

# Inspect the `NABMAb` assay data.
vrc01$datasets$NABMAb

# Load DAASH data from mab object
vrc01$loadDaash()

# Inspect DAASH datasets
vrc01$datasets$daash |> names()

}

}
\seealso{
\code{\link{connectDS}} \code{\link{DataSpaceConnection}}
}
\section{Super class}{
\code{\link[DataSpaceR:DataSpaceConnection]{DataSpaceConnection}} -> \code{DataSpaceMabs}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{mabMetadata}}{A data.table of mAbs with metadata found in
the object.}

    \item{\code{donorMetadata}}{A data.table of donors with metadata
found in the object.}

    \item{\code{mabMixMetadata}}{A data.table. A table of mAb mixtures
with metadata found in this DataSpaceMab instance.}

    \item{\code{mabMix}}{A data.table. A mapping table of mab_mix_id to mab_id.
with metadata found in this DataSpaceMab instance.}

    \item{\code{datasets}}{A list of data.table objects containing the
mab related that were loaded.}

    \item{\code{variableDefinitions}}{A data.table of variable definitions.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-DataSpaceMabs-initialize}{\code{DataSpaceMabs$new()}}
    \item \href{#method-DataSpaceMabs-print}{\code{DataSpaceMabs$print()}}
    \item \href{#method-DataSpaceMabs-loadDaash}{\code{DataSpaceMabs$loadDaash()}}
    \item \href{#method-DataSpaceMabs-refresh}{\code{DataSpaceMabs$refresh()}}
    \item \href{#method-DataSpaceMabs-clone}{\code{DataSpaceMabs$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="downloadPublicationData"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-downloadPublicationData'><code>DataSpaceConnection$downloadPublicationData()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="filterMabGrid"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-filterMabGrid'><code>DataSpaceConnection$filterMabGrid()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getDaash"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getDaash'><code>DataSpaceConnection$getDaash()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getDonors"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getDonors'><code>DataSpaceConnection$getDonors()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getGroup"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getGroup'><code>DataSpaceConnection$getGroup()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getGroups"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getGroups'><code>DataSpaceConnection$getGroups()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getMab"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getMab'><code>DataSpaceConnection$getMab()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getMabs"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getMabs'><code>DataSpaceConnection$getMabs()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getStudies"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getStudies'><code>DataSpaceConnection$getStudies()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getStudy"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getStudy'><code>DataSpaceConnection$getStudy()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="loadLanlMabMetadata"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-loadLanlMabMetadata'><code>DataSpaceConnection$loadLanlMabMetadata()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="resetMabGrid"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-resetMabGrid'><code>DataSpaceConnection$resetMabGrid()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceMabs-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceMabs-initialize}{}}}
\subsection{\code{DataSpaceMabs$new()}}{
  Initialize \code{DataSpaceMab} object.
See \code{\link{DataSpaceConnection}}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataSpaceMabs$new(mabIds, includeMixtures)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{mabIds}}{A character vector of `mab_id` values.}
      \item{\code{includeMixtures}}{Whether or not to include mab mixtures. "yes", "no", or "only" are valid.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceMabs-print"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceMabs-print}{}}}
\subsection{\code{DataSpaceMabs$print()}}{
  Print the \code{DataSpaceMab} object summary.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataSpaceMabs$print()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceMabs-loadDaash"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceMabs-loadDaash}{}}}
\subsection{\code{DataSpaceMabs$loadDaash()}}{
  Load any available DAASH datasets.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataSpaceMabs$loadDaash()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceMabs-refresh"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceMabs-refresh}{}}}
\subsection{\code{DataSpaceMabs$refresh()}}{
  Refresh the \code{DataSpaceMab} object to update datasets.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataSpaceMabs$refresh()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceMabs-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceMabs-clone}{}}}
\subsection{\code{DataSpaceMabs$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataSpaceMabs$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
