% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSpaceR.R
\name{DataSpaceR-package}
\alias{DataSpaceR-package}
\alias{DataSpaceR}
\title{A Thin Wrapper Around DataSpace}
\description{
DataSpaceR provides a convenient API for accessing data sets
within the DataSpace database.
}
\details{
Uses the Rlabkey package to connect to DataSpace. Implements
caching, and convenient methods for accessing data sets.
}
\author{
Ju Yeong Kim
}
