% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSpaceStudies.R
\name{DataSpaceStudies}
\alias{DataSpaceStudies}
\title{The DataSpaceStudies class}
\description{
An R6 class for DataSpace Study data.
}
\section{Constructor}{

\code{DataSpaceConnection$getStudies()}
}

\examples{
\dontrun{
# Create a connection (Initiate a DataSpaceConnection object)
con <- connectDS()

# Get group by `study_id` or pass a filtered `availableStudies` object.
studies <- con$getStudies(c("vtn505", "cvd408"))
studies <- con$getStudies(
  con$availableStudies[grepl("BAMA", data_availability) & species == "Human"]
)

# Load BAMA to the studies object.
studies$loadAssayDatasets("BAMA")
studies$datasets$BAMA

# Inspect variable information of the BAMA dataset
studies$datasetDescriptions$BAMA

# Inspect treatment arm information for all studies in study object
studies$treatmentArm

}

}
\seealso{
\code{\link{connectDS}} \code{\link{DataSpaceConnection}}
}
\section{Super class}{
\code{\link[DataSpaceR:DataSpaceConnection]{DataSpaceConnection}} -> \code{DataSpaceStudies}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{studies}}{A character vector of `study_id` values found in the object.}

    \item{\code{availableDatasets}}{A table of datasets available in
the \code{DataSpaceStudies} object.}

    \item{\code{datasets}}{A list of data.table objects containing the
availableDatasets that were loaded.}

    \item{\code{variableDefinitions}}{A list of data.table objects containing the
data dictionaries of the integrated data loaded.}

    \item{\code{treatmentArm}}{A data.table. The table of treatment arm
information for the connected study. Not available for all study
connection.}

    \item{\code{studyInfo}}{A list. Stores the information about the study.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-DataSpaceStudies-initialize}{\code{DataSpaceStudies$new()}}
    \item \href{#method-DataSpaceStudies-print}{\code{DataSpaceStudies$print()}}
    \item \href{#method-DataSpaceStudies-loadAvailableDatasets}{\code{DataSpaceStudies$loadAvailableDatasets()}}
    \item \href{#method-DataSpaceStudies-refresh}{\code{DataSpaceStudies$refresh()}}
    \item \href{#method-DataSpaceStudies-clone}{\code{DataSpaceStudies$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="downloadPublicationData"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-downloadPublicationData'><code>DataSpaceConnection$downloadPublicationData()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="filterMabGrid"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-filterMabGrid'><code>DataSpaceConnection$filterMabGrid()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getDaash"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getDaash'><code>DataSpaceConnection$getDaash()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getDonors"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getDonors'><code>DataSpaceConnection$getDonors()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getGroup"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getGroup'><code>DataSpaceConnection$getGroup()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getGroups"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getGroups'><code>DataSpaceConnection$getGroups()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getMab"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getMab'><code>DataSpaceConnection$getMab()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getMabs"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getMabs'><code>DataSpaceConnection$getMabs()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getStudies"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getStudies'><code>DataSpaceConnection$getStudies()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getStudy"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getStudy'><code>DataSpaceConnection$getStudy()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="loadLanlMabMetadata"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-loadLanlMabMetadata'><code>DataSpaceConnection$loadLanlMabMetadata()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="resetMabGrid"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-resetMabGrid'><code>DataSpaceConnection$resetMabGrid()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceStudies-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceStudies-initialize}{}}}
\subsection{\code{DataSpaceStudies$new()}}{
  Initialize \code{DataSpaceStudy} class.
See \code{\link{DataSpaceConnection}}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataSpaceStudies$new(studyIds)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{studyIds}}{A character. Name of the study to retrieve.
as URL, path and username.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceStudies-print"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceStudies-print}{}}}
\subsection{\code{DataSpaceStudies$print()}}{
  Print \code{DataSpaceStudy} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataSpaceStudies$print()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceStudies-loadAvailableDatasets"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceStudies-loadAvailableDatasets}{}}}
\subsection{\code{DataSpaceStudies$loadAvailableDatasets()}}{
  Load datasets to the studies object from an availableDatasets object.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataSpaceStudies$loadAvailableDatasets(
  availableDatasets = self$availableDatasets,
  downloadDir = tempdir()
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{availableDatasets}}{An `availableDatasets` object or vector of `study_id` values.}
      \item{\code{downloadDir}}{Optional, a character path specifying a directory to download.
nonstandard datasets. The default is the working temp directory.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceStudies-refresh"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceStudies-refresh}{}}}
\subsection{\code{DataSpaceStudies$refresh()}}{
  Refresh the study object to update available datasets and treatment info.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataSpaceStudies$refresh()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceStudies-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceStudies-clone}{}}}
\subsection{\code{DataSpaceStudies$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DataSpaceStudies$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
