% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSpaceStudy.R
\docType{class}
\name{DataSpaceStudy}
\alias{DataSpaceStudy}
\title{The DataSpaceStudy class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataSpaceStudy
}
\description{
The DataSpaceStudy class
}
\details{

}
\section{Constructor}{

\code{DataSpaceConnection$getStudy()}
}

\section{Fields}{

\describe{
  \item{\code{study}}{
    A character. The study name.
  }
  \item{\code{config}}{
    A list. Stores configuration of the connection object such as
    URL, path and username.
  }
  \item{\code{availableDatasets}}{
    A data.frame. The table of datasets available in the study object.
  }
  \item{\code{cache}}{
    A list. Stores the data to avoid downloading the same tables multiple
    times.
  }
  \item{\code{treatmentArm}}{
    A data.frame. The table of treatment arm information for the connected
    study. Not available for cross study connection.
  }
  \item{\code{group}}{
    A character. The group name.
  }
}
}

\section{Methods}{

\describe{
  \item{\code{initialize(study = NULL, config = NULL, group = NULL)}}{
    Initialize \code{DataSpaceStudy} class.
    See \code{\link{DataSpaceConnection}}.
  }
  \item{\code{print()}}{
    Print \code{DataSpaceStudy} class.
  }
  \item{\code{getDataset(datasetName, colFilter = NULL,
  reload = FALSE, ...)}}{
    Get a dataset from the connection.

    \code{datasetName}: A character. Name of the dataset to retrieve.

    \code{colFilter}: A matrix. A filter as returned by Rlabkey's
    \code{\link[Rlabkey]{makeFilter}}.

    \code{reload}: A logical. If set to TRUE, download the dataset, whether
    a cached version exist or not.

    \code{...}: Extra arguments to be passed to
    \code{\link[Rlabkey]{labkey.selectRows}}
  }
  \item{\code{clearCache()}}{
    Clear \code{cache}. Remove downloaded datasets.
  }
  \item{\code{getVariableInfo(datasetName)}}{
    Get variable information.

    \code{datasetName}: A character. Name of the dataset to retrieve.
  }
  \item{\code{refresh()}}{
    Refresh \code{DataSpaceStudy} class.
  }
}
}

\examples{
\dontrun{
# Create a connection (Initiate a DataSpaceConnection object)
con <- connectDS()

# Connect to cvd408 (Initiate a DataSpaceStudy object)
cvd408 <- con$getStudy("cvd408")

# Retrieve Neutralizing antibody dataset (NAb) for cvd408 from DataSpace
cvd408$getDataset("NAb")
}
}
\seealso{
\code{\link{connectDS}} \code{\link{DataSpaceConnection}}
}
\keyword{datasets}
