% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netrc.R
\name{write_netrc}
\alias{write_netrc}
\title{Write a netrc file}
\usage{
write_netrc(login, password, machine = "dataspace-staging.cavd.org",
  netrcFile = NULL)
}
\arguments{
\item{login}{A character. Email address used for logging in on DataSpace.}

\item{password}{A character. Password associated with the login.}

\item{machine}{A character. Host of DataSpace.}

\item{netrcFile}{A character. Credentials will be written into that file.
If left NULL, netrc will be written into a temporary file.}
}
\value{
A character vector containing the file paths for netrc
}
\description{
Write a netrc file that is valid for accessing DataSpace.
}
\examples{
\dontrun{
write_netrc("dataspaceuser@email.com", "mypassword")
}
}
\seealso{
\code{\link{connectDS}} \code{\link{check_netrc}}
}
