% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netrc.R
\name{checkNetrc}
\alias{checkNetrc}
\title{Check netrc file}
\usage{
checkNetrc(netrcFile = getNetrcPath(), onStaging = FALSE, verbose = TRUE)
}
\arguments{
\item{netrcFile}{A character. File path to netrc file to check.}

\item{onStaging}{A logical. Whether to check the staging server instead
of the production server.}

\item{verbose}{A logical. Whether to print the extra details for
troubleshooting.}
}
\value{
The name of the netrc file
}
\description{
Check that there is a netrc file with a valid entry for the
CAVD DataSpace.
}
\examples{
try(checkNetrc())
}
\seealso{
\code{\link{connectDS}} \code{\link{writeNetrc}}
}
