% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSpaceConnection.R
\docType{class}
\name{DataSpaceConnection}
\alias{DataSpaceConnection}
\title{The DataSpaceConnection class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataSpaceConnection
}
\description{
The DataSpaceConnection class
}
\details{

}
\section{Constructor}{

\code{\link{connectDS}}
}

\section{Fields}{

\describe{
  \item{\code{config}}{
    A list. Stores configuration of the connection object such as
    URL, path and username.
  }
  \item{\code{availableStudies}}{
    A data frame. The table of studies available in the connection object.
  }
}
}

\section{Methods}{

\describe{
  \item{\code{initialize(login = NULL, password = NULL, verbose = FALSE,
  onStaging = FALSE)}}{
    Initialize \code{DataSpaceConnection} class. See \code{\link{connectDS}}.
  }
  \item{\code{print()}}{
    Print \code{DataSpaceConnection} class.
  }
  \item{\code{getStudy(study)}}{
    Create a \code{DataSpaceStudy} class.

    \code{study}: A character. Name of the study to retrieve.
  }
}
}

\examples{
\dontrun{
# Create a connection (Initiate a DataSpaceConnection object)
con <- connectDS()
}
}
\seealso{
\code{\link{connectDS}} \code{\link{DataSpaceR-package}}
}
\keyword{datasets}
