% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSpaceGroups.R
\name{DataSpaceGroups}
\alias{DataSpaceGroups}
\title{The DataSpaceGroups class}
\description{
An R6 class for DataSpace Groups data.
}
\section{Constructor}{

\code{DataSpaceConnection$getGroups()}
}

\examples{
\dontrun{
# Create a connection (Initiate a DataSpaceConnection object)
con <- connectDS()

# Get group by `group_id` or pass a filtered `availableGroups` object.
groups <- con$getGroups(c(266, 267))
groups <- con$availableGroups[label == "NYVAC durability comparison"] |>
  con$getGroups()

# Retrieving group assay data for cvd408 from
# DataSpace is done automatically when the groups object is
# created.
groups$datasets$BAMA

# Get variable information of the assay dataset
groups$datasetDescription$BAMA

}

}
\seealso{
\code{\link{connectDS}} \code{\link{DataSpaceConnection}}
}
\section{Super class}{
\code{\link[DataSpaceR:DataSpaceConnection]{DataSpaceR::DataSpaceConnection}} -> \code{DataSpaceGroups}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{availableDatasets}}{A data.table of datasets available in
the object.}

\item{\code{datasets}}{A list of data.table objects containing the
availableDatasets that were loaded.}

\item{\code{variableDefinitions}}{A data.table of variable definitions.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataSpaceGroups-new}{\code{DataSpaceGroups$new()}}
\item \href{#method-DataSpaceGroups-print}{\code{DataSpaceGroups$print()}}
\item \href{#method-DataSpaceGroups-refresh}{\code{DataSpaceGroups$refresh()}}
\item \href{#method-DataSpaceGroups-clone}{\code{DataSpaceGroups$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="downloadPublicationData"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-downloadPublicationData'><code>DataSpaceR::DataSpaceConnection$downloadPublicationData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="filterMabGrid"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-filterMabGrid'><code>DataSpaceR::DataSpaceConnection$filterMabGrid()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getDaash"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getDaash'><code>DataSpaceR::DataSpaceConnection$getDaash()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getDonors"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getDonors'><code>DataSpaceR::DataSpaceConnection$getDonors()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getGroup"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getGroup'><code>DataSpaceR::DataSpaceConnection$getGroup()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getGroups"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getGroups'><code>DataSpaceR::DataSpaceConnection$getGroups()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getMab"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getMab'><code>DataSpaceR::DataSpaceConnection$getMab()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getMabs"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getMabs'><code>DataSpaceR::DataSpaceConnection$getMabs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getStudies"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getStudies'><code>DataSpaceR::DataSpaceConnection$getStudies()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="getStudy"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-getStudy'><code>DataSpaceR::DataSpaceConnection$getStudy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DataSpaceR" data-topic="DataSpaceConnection" data-id="resetMabGrid"><a href='../../DataSpaceR/html/DataSpaceConnection.html#method-DataSpaceConnection-resetMabGrid'><code>DataSpaceR::DataSpaceConnection$resetMabGrid()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceGroups-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceGroups-new}{}}}
\subsection{Method \code{new()}}{
Initialize `DataSpaceGroups` class.
See \code{\link{DataSpaceConnection}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSpaceGroups$new(groupIds = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groupIds}}{A character vecotor of `group_id` values.
as URL, path and username.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceGroups-print"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceGroups-print}{}}}
\subsection{Method \code{print()}}{
Print \code{DataSpaceStudy} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSpaceGroups$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceGroups-refresh"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceGroups-refresh}{}}}
\subsection{Method \code{refresh()}}{
Refresh loaded integrated datasets, and information of what datasets are available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSpaceGroups$refresh()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSpaceGroups-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataSpaceGroups-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSpaceGroups$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
