% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSpaceR.R
\name{DataSpaceR-package}
\alias{DataSpaceR-package}
\alias{DataSpaceR}
\title{DataSpaceR}
\description{
DataSpaceR provides a convenient API for accessing datasets
within the DataSpace database.
}
\details{
Uses the Rlabkey package to connect to DataSpace. Implements
convenient methods for accessing datasets.
}
\seealso{
\code{\link{connectDS}}
}
\author{
Ju Yeong Kim
}
