% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netrc.R
\name{write_netrc}
\alias{write_netrc}
\title{Write a netrc file}
\usage{
write_netrc(login, password, file = NULL)
}
\arguments{
\item{login}{A character. The email address used for loging in on DataSpace.}

\item{password}{A character. The password associated with the login.}

\item{file}{A character. The credentials will be written into that file.
If left NULL, the netrc will be written into a temporary file.}
}
\value{
A character vector containing the file paths for netrc
}
\description{
Write a netrc file that is valid for accessing DataSpace
}
\examples{
\dontrun{
write_netrc("dataspaceuser@email.com", "mypassword")
}
}
\seealso{
\code{\link{CreateConnection}} \code{\link{check_netrc}}
}
