% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netrc.R
\name{writeNetrc}
\alias{writeNetrc}
\title{Write a netrc file}
\usage{
writeNetrc(login, password, onStaging = FALSE,
  netrcFile = getNetrcPath())
}
\arguments{
\item{login}{A character. Email address used for logging in on DataSpace.}

\item{password}{A character. Password associated with the login.}

\item{onStaging}{A logical. Whether to connect to the staging server instead
of the production server.}

\item{netrcFile}{A character. Credentials will be written into that file.
If left NULL, netrc will be written into a temporary file.}
}
\value{
A character vector containing the file paths for netrc
}
\description{
Write a netrc file that is valid for accessing DataSpace.
}
\examples{
\dontrun{
writeNetrc("dataspaceuser@email.com", "mypassword")
}
}
\seealso{
\code{\link{connectDS}} \code{\link{checkNetrc}}
}
